Good place to continue with today.
I have the ship components iteratively generating based on a list of numbers at the moment with them looking not bad just now. Maybe can get pictures of the ships when loading them. Now setting up calls to api with fetch api.

no idea if the output.css file is needed but going to push it up just in case.